# openapi.json
{"openapi":"3.1.0","info":{"title":"Portexa API","version":"0.1.0"},"paths":{"/api/v1/ai/chat":{"post":{"tags":["ai"],"summary":"Chat","operationId":"chat_api_v1_ai_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AIChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AIChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/ai/rag-chat":{"post":{"tags":["ai"],"summary":"Rag Chat","operationId":"rag_chat_api_v1_ai_rag_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AIChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AIChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/ai/structure-resume":{"post":{"tags":["ai"],"summary":"Structure Resume","operationId":"structure_resume_api_v1_ai_structure_resume_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResumeStructureRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResumeStructureResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/signup":{"post":{"tags":["auth"],"summary":"Signup","operationId":"signup_api_v1_auth_signup_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_api_v1_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLogin"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/google":{"post":{"tags":["auth"],"summary":"Google Login","description":"Google OAuth login endpoint.\nVerify Google ID token and create/login user.","operationId":"google_login_api_v1_auth_google_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GoogleTokenRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/me":{"get":{"tags":["auth"],"summary":"Me","operationId":"me_api_v1_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"put":{"tags":["auth"],"summary":"Update Me","operationId":"update_me_api_v1_auth_me_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileUpdateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/embed/customization/{portfolio_id}":{"get":{"tags":["embed"],"summary":"Get Embed Customization","description":"Get customization for public embed chat.\nPortfolio ID is used to fetch user and their customization.","operationId":"get_embed_customization_api_v1_embed_customization__portfolio_id__get","parameters":[{"name":"portfolio_id","in":"path","required":true,"schema":{"type":"string","title":"Portfolio Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/embed/rag-chat":{"post":{"tags":["embed"],"summary":"Embed Rag Chat","description":"Public RAG-grounded chat endpoint for embedded widget.\nUses portfolio_id to fetch user's active resume and customization.","operationId":"embed_rag_chat_api_v1_embed_rag_chat_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmbedChatRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmbedChatResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/health":{"get":{"tags":["health"],"summary":"Health Check","operationId":"health_check_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/resumes":{"get":{"tags":["resumes"],"summary":"Get Resumes","operationId":"get_resumes_api_v1_resumes_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResumeLibraryResponse"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["resumes"],"summary":"Add Resume","operationId":"add_resume_api_v1_resumes_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_add_resume_api_v1_resumes_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResumeLibraryItem"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/resumes/{entry_id}":{"delete":{"tags":["resumes"],"summary":"Remove Resume","operationId":"remove_resume_api_v1_resumes__entry_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"entry_id","in":"path","required":true,"schema":{"type":"string","title":"Entry Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/resumes/{entry_id}/activate":{"patch":{"tags":["resumes"],"summary":"Activate Saved Resume","operationId":"activate_saved_resume_api_v1_resumes__entry_id__activate_patch","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"entry_id","in":"path","required":true,"schema":{"type":"string","title":"Entry Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/resumes/customization":{"get":{"tags":["resumes"],"summary":"Get Customization","operationId":"get_customization_api_v1_resumes_customization_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomizationConfig"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"put":{"tags":["resumes"],"summary":"Update Customization","operationId":"update_customization_api_v1_resumes_customization_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomizationConfig"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CustomizationConfig"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"AIChatMessage":{"properties":{"role":{"type":"string","pattern":"^(system|user|assistant)$","title":"Role"},"content":{"type":"string","title":"Content"}},"type":"object","required":["role","content"],"title":"AIChatMessage"},"AIChatRequest":{"properties":{"prompt":{"type":"string","title":"Prompt"},"model":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model"},"messages":{"anyOf":[{"items":{"$ref":"#/components/schemas/AIChatMessage"},"type":"array"},{"type":"null"}],"title":"Messages"}},"type":"object","required":["prompt"],"title":"AIChatRequest"},"AIChatResponse":{"properties":{"response":{"type":"string","title":"Response"},"model":{"type":"string","title":"Model"}},"type":"object","required":["response","model"],"title":"AIChatResponse"},"Body_add_resume_api_v1_resumes_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"},"index_json":{"type":"string","title":"Index Json"}},"type":"object","required":["file","index_json"],"title":"Body_add_resume_api_v1_resumes_post"},"CustomizationConfig":{"properties":{"theme":{"type":"string","title":"Theme","default":"light"},"accent":{"type":"string","title":"Accent","default":"#6366f1"},"bubbleStyle":{"type":"string","title":"Bubblestyle","default":"rounded"},"showWelcomeMessage":{"type":"boolean","title":"Showwelcomemessage","default":true}},"type":"object","title":"CustomizationConfig"},"EmbedChatRequest":{"properties":{"message":{"type":"string","title":"Message"},"portfolio_id":{"type":"string","title":"Portfolio Id"}},"type":"object","required":["message","portfolio_id"],"title":"EmbedChatRequest"},"EmbedChatResponse":{"properties":{"response":{"type":"string","title":"Response"}},"type":"object","required":["response"],"title":"EmbedChatResponse"},"GoogleTokenRequest":{"properties":{"token":{"type":"string","title":"Token"}},"type":"object","required":["token"],"title":"GoogleTokenRequest","description":"Schema for Google OAuth token verification"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ProfileUpdateRequest":{"properties":{"full_name":{"type":"string","title":"Full Name"},"password":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Password"}},"type":"object","required":["full_name"],"title":"ProfileUpdateRequest"},"ResumeExperienceItem":{"properties":{"title":{"type":"string","title":"Title"},"company":{"type":"string","title":"Company"},"period":{"type":"string","title":"Period"}},"type":"object","required":["title","company","period"],"title":"ResumeExperienceItem"},"ResumeLibraryItem":{"properties":{"id":{"type":"string","title":"Id"},"file_name":{"type":"string","title":"File Name"},"source_type":{"type":"string","title":"Source Type"},"indexed_at":{"type":"string","format":"date-time","title":"Indexed At"},"name":{"type":"string","title":"Name"},"summary":{"type":"string","title":"Summary"},"experience":{"items":{"additionalProperties":{"type":"string"},"type":"object"},"type":"array","title":"Experience"},"education":{"items":{"type":"string"},"type":"array","title":"Education"},"skills":{"items":{"type":"string"},"type":"array","title":"Skills"},"raw_text":{"type":"string","title":"Raw Text"},"size_bytes":{"type":"integer","title":"Size Bytes"},"is_active":{"type":"boolean","title":"Is Active","default":false}},"type":"object","required":["id","file_name","source_type","indexed_at","name","summary","experience","education","skills","raw_text","size_bytes"],"title":"ResumeLibraryItem"},"ResumeLibraryResponse":{"properties":{"active_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Active Id"},"items":{"items":{"$ref":"#/components/schemas/ResumeLibraryItem"},"type":"array","title":"Items"}},"type":"object","required":["items"],"title":"ResumeLibraryResponse"},"ResumeStructureRequest":{"properties":{"file_name":{"type":"string","title":"File Name"},"raw_text":{"type":"string","title":"Raw Text"},"source_type":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source Type"}},"type":"object","required":["file_name","raw_text"],"title":"ResumeStructureRequest"},"ResumeStructureResponse":{"properties":{"file_name":{"type":"string","title":"File Name"},"source_type":{"type":"string","title":"Source Type"},"name":{"type":"string","title":"Name"},"summary":{"type":"string","title":"Summary"},"experience":{"items":{"$ref":"#/components/schemas/ResumeExperienceItem"},"type":"array","title":"Experience"},"education":{"items":{"type":"string"},"type":"array","title":"Education"},"skills":{"items":{"type":"string"},"type":"array","title":"Skills"},"raw_text":{"type":"string","title":"Raw Text"}},"type":"object","required":["file_name","source_type","name","summary","experience","education","skills","raw_text"],"title":"ResumeStructureResponse"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"TokenResponse"},"UserCreate":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"full_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Name"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserCreate"},"UserLogin":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"UserLogin"},"UserRead":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"full_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Name"},"id":{"type":"integer","title":"Id"},"is_active":{"type":"boolean","title":"Is Active"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["email","id","is_active","created_at"],"title":"UserRead"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/v1/auth/login"}}}}}}